scheduler: Fix TestMultipleHosts
It was previously using explicit job IDs to check host jobs but even
though we generate the IDs in a predictable way, StartJob is called in a
goroutine so the order with which jobs are started is not deterministic.

I have modified the test to just check that each host has the expected
jobs based on type and state, as that is what matters.

Signed-off-by: Lewis Marshall <lewis@lmars.net>